# docs/swagger.yml
openapi: '3.0.2'
info:
  title: RecyThing
  version: '1.0'
  description: "API Documentation Capstone Project RecyThing Kelompok 2"
servers:
  - url: http://localhost:8080/api/v1
...
paths:
...
  /article/{articleId}:
    get:
      tags:
        - articles
        - manage articles
      summary: Get article data by id
      description: Endpoint user to get article by id
      operationId: getArticleById
      parameters:
        - name: articleId
          in: query
          required: true
          schema:
            type: string
            # default: ART0001
          description: Article ID
          example: ART0001
      security:
        - Bearer: []
      responses:
        '200':
          description: Get data article
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message: 
                    type: string
                    example: ok
                  data:
                    type: object
                    properties:
                      schema:
                        $ref: '#/components/schemas/Article'
        '400':
          description: Invalid param request
        '404':
          description: Article not found
    put:
      tags:
        - manage articles
      summary: Edit article
      description: Endpoint admin to edit an existing article.
      operationId: editArticle
      parameters:
        - name: articleId
          in: path
          required: true
          schema:
            type: string
          description: ID of the article to be edited
          example: ART0001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Article'
      security:
        - Bearer: []
      responses:
        '200':
          description: Article successfully updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: article updated!
        '404':
          description: Article not found
        '400':
          description: Invalid request data 
    delete:
      tags:
        - manage articles
      summary: Delete article
      description: Endpoint admin to delete an article.
      operationId: deleteArticle
      parameters:
        - name: articleId
          in: path
          required: true
          schema:
            type: string
          description: ID of the article to be deleted
          example: ART0001
      security:
        - Bearer: []
      responses:
        '200':
          description: Article successfully deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: article successfully deleted!
        '404':
          description: Article not found
...
  /videos/categories:
    get:
      tags:
        - manage videos
      summary: Get All Categories
      description: Get All Categories
      operationId: getAllCategories
      security:
        - Bearer: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        content_categories:
                          type: array
                          items:
                            type: object
                            properties:
                              name:
                                type: string
                                example: "tips"
                        waste_categories:
                          type: array
                          items:
                            type: object
                            properties:
                              name:
                                type: string
                                example: "berbahaya"
                      
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: error get all categories
...
components:
  schemas:
...
    Article:
      type: object
      properties:
        id:
          type: string
          example: ART0001
        author:
          type: object
          properties:
            id:
              type: string
              example: AD0001
            name: 
              type: string
              example: John Doe
            image_url:
              type: string
              example: http://example.com/image.png
        title:
          type: string
          example: Bahaya Limbah Elektronik dan Cara Pembuangan yang Bertanggung Jawab
        description:
          type: string
          example: Pelajari tentang bahaya lingkungan dan kesehatan dari limbah elektronik dan temukan metode pembuangan yang aman.
        thumbnail_url:
          type: string
          format: uri
          example: http://example.com/image.png
        waste_categories:
          type: array
          items:
            type: string
          example: ["elektronik", "berbahaya"]
        content_categories:
          type: array
          items:
            type: string
          example: ["edukasi", "kampanye"]
        sections:
          type: array
          items:
            $ref: '#/components/schemas/ArticleSection'
    ArticleSection:
      type: object
      properties:
        title:
          type: string
          example: Apa itu Limbah Elektronik?
        description:
          type: string
          example: Limbah elektronik mencakup barang elektronik bekas seperti ponsel, komputer, peralatan rumah tangga, dll.
        image_url:
          type: string
          format: uri
          example: http://example.com/image.png
...